Support integer values for numbers beyond the range of 32 bit integers but within the RFC allowed 64 bit integers.

+++ src/DotLiquid/Tags/For.cs
@@ -107,8 +107,8 @@ namespace DotLiquid.Tags
                     : Convert.ToInt32(context[_attributes["offset"]])
                 : 0;
 
-            int? limit = _attributes.ContainsKey("limit") ? context[_attributes["limit"]] as int? : null;
-            int? to = (limit != null) ? (int?) (limit.Value + from) : null;
+            long? limit = _attributes.ContainsKey("limit") ? context[_attributes["limit"]] as long? : null;
+            long? to = (limit != null) ? (long?) (limit.Value + from) : null;
 
             List<object> segment = SliceCollectionUsingEach(context, (IEnumerable) collection, from, to);
 
@@ -166,7 +166,7 @@ namespace DotLiquid.Tags
             });
         }
 
-        private static List<object> SliceCollectionUsingEach(Context context, IEnumerable collection, int from, int? to)
+        private static List<object> SliceCollectionUsingEach(Context context, IEnumerable collection, int from, long? to)
         {
             List<object> segments = new List<object>();
             int index = 0;
